Reorganized Files & Fixed Bugs
Fixed some bugs and warnings
Made the file structure more organized for the finished product of this semester.
