+++ source/appCore/keyboardController.h
@@ -1,0 +1,11 @@
+#ifndef KEYBOARD_CONTROLLER_H
+#define KEYBOARD_CONTROLLER_H
+
+namespace JCAT {
+    class KeyboardController {
+        public:
+        private:
+    };
+};
+
+#endif
