+++ source/engine/2d/src/camera2D.cpp
@@ -1,0 +1,8 @@
+#include <cassert>
+#include <limits>
+
+#include "./engine/2d/camera2D.h"
+
+namespace JCAT {
+
+};
