+++ source/engine/deviceSetup.h
@@ -18,7 +18,10 @@
 
 /** Used for checking if the current device is plugged in or not (Imports based on operating system) */
 #if defined(_WIN32) || defined(_WIN64)
-    #define NOMINMAX
+    #ifndef NOMINMAX
+        #define NOMINMAX
+    #endif
+    
     #include <windows.h>
 #elif defined(__linux__)
     #include <fstream>
